requirements.txt
Flask~=3.1.0
Flask-MySQL
pandas~=2.2.3
numpy~=2.0.2
matplotlib
scikit-learn
mysql-connector-python~=9.1.0
email-validator
smtplib
os

shap~=0.46.0
geopy~=2.4.1